Move compile.js, helpers.js, connection.js to src/util Since they are utilities that are used in multiple parts of the whole package.
