+++ src/transport/DataConnection.js
@@ -1,6 +1,6 @@
 const log = require('debug')('bilibili-danmaku-client/DataConnection');
 
-const { CascadeConnection } = require('../connection');
+const { CascadeConnection } = require('../util/connection');
 const { SectionConnection, Section, SectionCoder, StringCoder, JsonCoder } = require('./SectionConnection');
 
 const handshakeCoder = new JsonCoder({ controlFlag: true, opCode: 7, binaryFlag: true });
